The Login System IMPLEMENTED!
This has taken over ~1 day straight (24 hours) of work to implement this bug free working login system. The login system features token validation and more.
Error handling such as 413 is to come as well as useable actual features in the next update. Pretty much the same code will be used in StreamBop and SnailPortal.
FEATURES THAT NEED TO BE ADDED:
- Error HANDLING. :sweat:
- STABILITY.
- More CHECKS.
- Better language responses
SECURITY. and such as performance.